style(dataSource): code style
