+++ packages/rath-client/src/pages/dataConnection/index.tsx
@@ -17,6 +17,7 @@ import React, { useCallback, useState } from 'react';
 import { Breadcrumb, DefaultButton, IBreadcrumbItem, ProgressIndicator } from '@fluentui/react';
 import { observer } from 'mobx-react-lite';
 import intl from 'react-intl-universal';
+import styled from 'styled-components';
 import { IDataSourceType } from '../../global';
 import { IMuteFieldBase, IRow } from '../../interfaces';
 import { DataSourceTag, IDBMeta, setDataStorage } from '../../utils/storage';
@@ -45,6 +46,20 @@ interface DataConnectionProps {
     // onDataLoading: (p: number) => void;
     // toggleLoadingAnimation: (on: boolean) => void;
 }
+
+const FormContainer = styled.div`
+    flex-grow: 1;
+    flex-shrink: 0;
+    flex-basis: max-content;
+    display: flex;
+    flex-direction: column;
+`;
+
+const FirstClassMenu = styled.div`
+    flex-grow: 9999;
+    flex-shrink: 0;
+    flex-basis: max-content;
+`;
 
 const DataConnection: React.FC<DataConnectionProps> = (props) => {
     const { dataSourceStore, commonStore } = useGlobalStore();
@@ -171,20 +186,20 @@ const DataConnection: React.FC<DataConnectionProps> = (props) => {
                         />
                     </div>
                 )}
-                <div style={{ flexGrow: 1, flexShrink: 0, flexBasis: 'max-content', flexDirection: 'column' }}>
+                <FormContainer>
                     {loading && dataSourceType !== IDataSourceType.FILE && <ProgressIndicator description="loading" />}
                     {loading && dataSourceType === IDataSourceType.FILE && <DataLoadingStatus />}
                     {dataSourceType && formMap[dataSourceType]}
-                </div>
+                </FormContainer>
             </Card>
             {dataSourceType === null && (
-                <div style={{ flexGrow: 9999, flexShrink: 0, flexBasis: 'max-content' }}>
+                <FirstClassMenu>
                     <SupportedSources
                         onSelected={(k) => {
                             setDataSourceType(k as IDataSourceType);
                         }}
                     />
-                </div>
+                </FirstClassMenu>
             )}
         </div>
     );
