+++ packages/rath-client/src/pages/dataSource/selection/database/form/connect-options.tsx
@@ -1,5 +1,5 @@
 import intl from 'react-intl-universal';
-import { FC, useEffect, useMemo, useState } from 'react';
+import { FC, ComponentPropsWithRef, useEffect, useMemo, useState } from 'react';
 import { observer } from 'mobx-react-lite';
 import styled from 'styled-components';
 import { DefaultButton, Dropdown, IDropdownOption, Label, Stack, TextField } from '@fluentui/react';
@@ -13,13 +13,50 @@ const Form = styled.div`
     grid-template-columns: max-content 1fr;
     column-gap: 0.5em;
     row-gap: 0.4em;
-    & :nth-child(3n) {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
+`;
+
+const FlexingTextField = styled(TextField)`
+    flex-grow: 1;
+    flex-shrink: 1;
+    .ms-TextField-fieldGroup {
+        border-radius: 0 4px 4px 0;
     }
 `;
+
+const ConnectUriField = styled(FlexingTextField)`
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+    .ms-TextField-wrapper {
+        width: 100%;
+    }
+    .ms-TextField-fieldGroup {
+        border-radius: 0 4px 4px 0;
+    }
+`;
+
+const DatabaseDropdownStyles: ComponentPropsWithRef<typeof Dropdown>['styles'] = {
+    dropdown: {
+        width: '13.6em',
+        borderRadius: '2px 0 0 2px',
+    },
+    dropdownItems: {
+        paddingBlockStart: '6px',
+        paddingBlockEnd: '6px',
+        maxHeight: '20vh',
+        overflowY: 'scroll',
+    },
+    dropdownItemSelected: {
+        position: 'static',
+        minHeight: '2.2em',
+    },
+    dropdownItem: {
+        position: 'static',
+        minHeight: '2.2em',
+    },
+};
 
 interface ConnectOptionsProps {
     disabled: boolean;
@@ -72,26 +109,7 @@ const ConnectOptions: FC<ConnectOptionsProps> = ({
                         ariaLabel={intl.get('dataSource.databaseType')}
                         disabled={disabled}
                         required
-                        styles={{
-                            dropdown: {
-                                width: '13.6em',
-                                borderRadius: '2px 0 0 2px',
-                            },
-                            dropdownItems: {
-                                paddingBlockStart: '6px',
-                                paddingBlockEnd: '6px',
-                                maxHeight: '20vh',
-                                overflowY: 'scroll',
-                            },
-                            dropdownItemSelected: {
-                                position: 'static',
-                                minHeight: '2.2em',
-                            },
-                            dropdownItem: {
-                                position: 'static',
-                                minHeight: '2.2em',
-                            },
-                        }}
+                        styles={DatabaseDropdownStyles}
                         options={databaseOptions}
                         selectedKey={sourceType}
                         onRenderOption={renderDropdownItem as (e?: IDropdownOption) => JSX.Element}
@@ -102,7 +120,7 @@ const ConnectOptions: FC<ConnectOptionsProps> = ({
                             }
                         }}
                     />
-                    <TextField
+                    <ConnectUriField
                         name={databaseConfig ? `connectUri:${databaseConfig.key}` : undefined}
                         title={intl.get('dataSource.connectUri')}
                         aria-required
@@ -110,23 +128,12 @@ const ConnectOptions: FC<ConnectOptionsProps> = ({
                         value={connectUri}
                         placeholder={databaseConfig?.rule}
                         onChange={(_, uri) => setConnectUri(uri ?? '')}
-                        styles={{
-                            root: {
-                                position: 'relative',
-                                flexGrow: 1,
-                                flexShrink: 1,
-                            },
-                            fieldGroup: {
-                                borderLeft: 'none',
-                                borderRadius: '0 4px 4px 0',
-                            },
-                        }}
                     />
                 </Stack>
             </Form>
             {databaseConfig?.credentials === 'json' && (
                 <Stack horizontal verticalAlign="end">
-                    <TextField
+                    <FlexingTextField
                         multiline
                         autoComplete="false"
                         title="Credentials"
@@ -136,15 +143,6 @@ const ConnectOptions: FC<ConnectOptionsProps> = ({
                         value={credentialsRaw}
                         placeholder="{}"
                         onChange={(_, content) => setCredentialsRaw(content ?? '')}
-                        styles={{
-                            root: {
-                                flexGrow: 1,
-                                flexShrink: 1,
-                            },
-                            fieldGroup: {
-                                borderRadius: '0 4px 4px 0',
-                            },
-                        }}
                     />
                     <DefaultButton
                         text={intl.get('common.apply')}
